[T] test: move test component
