+++ __tests__/components/discover/roadmap-detail/view/timeLineView.test.js
@@ -1,7 +1,7 @@
 import 'react-native';
 import React, {} from 'react';
 import renderer from 'react-test-renderer';
-import TimeLineView from '../../../src/components/discover/toadmap-detail/view/TimeLineView';
+import TimeLineView from '../../../../../src/components/discover/roadmap-detail/view/TimeLineView';
 
 
 it('renders correctly', () => {
