+++ __tests__/components/discover/view/discoverMoreItem.test.js
@@ -1,9 +1,9 @@
 import 'react-native';
 import React, {} from 'react';
 import renderer from 'react-test-renderer';
-import DiscoverMoreItem from '../../../src/components/discover/view/DiscoverMoreItem';
+import DiscoverMoreItem from '../../../../src/components/discover/view/DiscoverMoreItem';
 
-const home3 = require('../../../assets/growth-ui/img/home-3.jpg');
+const home3 = require('../../../../assets/growth-ui/img/home-3.jpg');
 
 it('renders correctly', () => {
   const moreItem = renderer.create(
