+++ __tests__/components/discover/view/discoverOnlineReadItem.test.js
@@ -1,9 +1,9 @@
 import 'react-native';
 import React, {} from 'react';
 import renderer from 'react-test-renderer';
-import DiscoverOnlineReadItem from '../../../src/components/discover/view/DiscoverOnlineReadItem';
+import DiscoverOnlineReadItem from '../../../../src/components/discover/view/DiscoverOnlineReadItem';
 
-const home1 = require('../../../assets/growth-ui/img/home-1.jpg');
+const home1 = require('../../../../assets/growth-ui/img/home-1.jpg');
 
 it('renders correctly', () => {
   const onlinereadItem = renderer.create(
